# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de proveedores",
        "description": "Documentación de la API para gestión de proveedores y Productoos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/catalogos": {
            "get": {
                "tags": [
                    "Catálogos"
                ],
                "summary": "Listar catálogos",
                "operationId": "6be93f181b751655105df230ac037ef8",
                "responses": {
                    "200": {
                        "description": "Lista de catálogos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Catalogo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Catálogos"
                ],
                "summary": "Crear catálogo",
                "operationId": "26d350396d7e86077e89a3051c9ac841",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Catalogo"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catálogo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Catalogo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalogos/{id}": {
            "get": {
                "tags": [
                    "Catálogos"
                ],
                "summary": "Obtener catálogo por ID",
                "operationId": "e9b85c5025015d6425b875eff1b20f91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catálogo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Catalogo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catálogos"
                ],
                "summary": "Actualizar catálogo",
                "operationId": "548b5db253cf82f5cbcb889c9d613d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Catalogo"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catálogo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Catalogo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catálogos"
                ],
                "summary": "Eliminar catálogo",
                "operationId": "f6f3e29c671ae9ff17f347e63aab2aec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Catálogo eliminado"
                    }
                }
            }
        },
        "/api/categoria": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Listar categorías",
                "operationId": "39a25a0c4ca44e7a15bdca633a64de07",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Categoria"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Crear categoría",
                "operationId": "31caf87e28eaca6c42c1833d9f03a536",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categoria/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener categoría por ID",
                "operationId": "cc96c4e2c9d382a6312b612182a613a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar categoría",
                "operationId": "b705c18d15454cf322ab683013568dda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Categoria"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar categoría",
                "operationId": "dff942a16b1b0ff279264e8f1a19b942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoría eliminada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Catalogo": {
                "required": [
                    "nombre",
                    "proveedor_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Catálogo 2024"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Catálogo anual de Productoos",
                        "nullable": true
                    },
                    "proveedor_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/catalogos/2024.jpg",
                        "nullable": true
                    },
                    "Proveedor": {
                        "$ref": "#/components/schemas/Proveedor"
                    }
                },
                "type": "object"
            },
            "Categoria": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Herramientas eléctricas"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Categoría de herramientas eléctricas",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/categoria/electricas.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Proveedor": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Suministros Industriales S.A."
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Central 123",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-1234",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "contacto@proveedor.com",
                        "nullable": true
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/proveedores/si.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}